requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
